refactor: tableViewCell -> CollectionViewCell 로 변경 및 TableView메서드 제거
